Bugfix Select All Button Slightly Too Large on Mobile
The select all button on mobile was slightly too big due to a top margin
on the input element which caused the element to extend in height
slightly.

Positioning has also been added to fix some vertical alignment of the
input element within the label element as the alignment was slightly off
on both larger and smaller screen sizes.