Made source and licence open in tab
```
source_link = "<a href='#{Octobox.config.source_repo}' target='_blank'>Source</a>"
    license_link = "<a href='#{Octobox.config.source_repo}/blob/master/LICENSE.txt' target='_blank'>AGPL 3.0</a>"
```

I added`target='_blank'` to the source and license link. I believe this is the way to do it, but note the standard way is `target="_blank"`. I did it with `'` as that is how the links are added, and using `"` would cause a conflict.

Thanks, hope this helps!