Feature: Display author of notification subject
When notifications are synced, the author of the subject of the
notification is recorded by making an call to the subject (issue,
commit, whatever) using the notification response.

This is only done once per notification, as the author should never
change and making an extra API per notification on sync would be
terrible.

Displays the author in the main notifications table.

Prefixes author with 'by' for mobile