Standardized README (Mostly)
So, this is actually already an amazing README. I really love it. This means I didn't standardize all of it - for instance, I didn't put the description in a quote block beneath the badges, because your screenshot is fine. I also kept the Why Is This A Thing header, and Getting Started. However, I did move some stuff around. For instance, backers and supporters should be on the bottom. They're not really relevant to people who come to this README for the first time. I also added a Table of Contents, and filled out the Contribute section a bit more - Development wasn't as clear as it could have been. I added a note about Help Wanted labels there, too, and I included the name of the license in the Copyright section.

Other than those small things, this looks great!