Fix json notification response when zero (#395)
A divide by zero exception occurs when no notifications are returned in json. Adds a test to verify the pagination attributes of the json response for 1 or more notifications as well as zero notifications. Makes total_pages use the @per_page instance variable instead of @cur_selected, to avoid the divide by zero issue.