Download Subjects v2 (#411)
* Work in progress on #4

* Feature: Display author of notification subject

When notifications are synced, the author of the subject of the
notification is recorded by making an call to the subject (issue,
commit, whatever) using the notification response.

This is only done once per notification, as the author should never
change and making an extra API per notification on sync would be
terrible.

Displays the author in the main notifications table.

Prefixes author with 'by' for mobile

* Fix non-existant local var without capture

Using =~ doesn't work here as if there's no issue_number capture the
local var isn't even initialized. Oops.

* Warn on unknown notification subjects

I don't want to raise here as that allows any changes to Github to bring
the app down until it's resolved. Instead, just don't set the author and
put a warning in the logs.

* Recognize release authors

* Store subjects in a seperate table

* Subjects should be optional

* Remove unused regex

* Refactored #update_subject to reduce excess http requests

* Add feature flag to #update_subject

* Subjects can have many notifications

* Store subject timestamps when syncing

* Only sync subjects if notification is newer than the existing subject

* Eager load subjects when syncing notifications

* Add index on subject url

* Allow nil when delegating to subjects

* Tests for ensuring subjects get created

* More intuitive time-based subject skip w/tests

* Test merged PRs get updated correctly

* Fix unrelated test

Since I added another notification to this fixture

* Whitespace

* Add documentation for download subjects feature flag

* Added rake task for syncing missing subjects

* Update subjects in batches from rake task

* Fix ambiguous column name in newest scope