Updated Web App Manifest (#427)
The current Web App Manifest was not really populated with usable data. This commits adds recommended fields, and includes a range of icon sizes to fit a variety of common use cases.

If accessed on a Chromebook, or on an Android device, and then add the application to the drawer or home screen respectively, it'll look behave and work much nicer.