Link to last comment (#426)
* Add latest_comment_url to notifications

* Sync notifications' latest_comment_url and use it

* Generalize to a URL parser, include commitcomments

* Account for always-populated latest_comment_url

GitHub apparently always populates the subject's latest_comment_url,
defaulting to the subject in general if there's no specific comment to
link to. Also it resets back to the subject on a major state change like
a merge. So, only add the anchor if the latest_comment_url is explicitly
a comment.

* Sprinkle a little DI on SubjectUrlParser

Adds ability to inject whatever domains you want to parse with,
defaulting to the application configuration.

* Add some tests for SubjectUrlParser