Allow API requests past protect_from_forgery (#455)
Allows API requests a method to bypass the CSRF protection. Adding a custom header that disables the protection makes it imposible for a browser to disable the CSRF protection but easy for the application using the API to opt out. Signed cookies are still used for authentication, so the API requests still need to authenticate first and pass the user_id.