Show loading screen on initial notification sync
When signing up as a new user, you'll currently stare at a blank screen
while the OAuth callback endpoint is getting hit and your notifications
are being synced for the first time. For users with a huge number of
notifications, this can take a *long* time (multiple minutes) without
any feedback to the user at all. It also usually times out and gives a
confusing auth error, leaving the user wondering what happened.

These changes make an exception to the initial sync process, instead
loading the page and triggering a sync event with a loading screen to
assure the user that things are going and it'll take a bit.