Fix escape on help modal triggering page reload
The modal is a bootstrap modal, which helpfully binds escape to
the action of closing the modal. People who are learning how to
use Octobox for the first time might be selecting elements, then
hitting ? and reading the shortcuts, then escape, only to find
their previous selection erased due to the page reload.