chore(docker, openshift): Add Docker build for octobox + sidekiq
Added a new "Dockerfile-with-sidekiq" which uses foreman to run
two processes: rails with octobox and a sidekiq worker.
Althought this conflicts with the 'pid 1' paradigm to only run a single
process within an Docker container, it saves some resources when
compared when used in two separate containers.

This is especially useful in environments which enforce a minimum size
for heap memory like OpenShift Online, which in its "Starter tier" also
has only a limited budget for memory.