feature(docker): Adapt Docker image to run sidekiq when enabled (#943)
* feature(docker): Adapt Docker image to run sidekiq when enabled

Modified Docker image to start check OCTOBOX_SIDEKIQ_SCHEDULE_ENABLED
and start foreman with rails+sidekiq if set. If not set, then
rails is start as usual.

* chore(docker): Update startup script to be more lenient

* chore(docker): Optimized Dockerfile

The instructions are collected into two groups/layers:

* The parts for the dependencies (linux, ruby) goes first, which means
that as long as the GemFile doesn't change this layer is cached. Note,
that the OS will be also only updated if the the Gemfile changes.
Updating OS packages (with an unconditional --update) this way
also leads to not reproducible images as it depends _when_ you build
the image. We should consider pinning the apk versions, too.

* The second part contains the application source and will run always
if something changes in the directory (nearly always the case)

This is a good compromise between short turn around times and a proper
caching.