Prompt users to log into the GitHub App
If users have notifications from repositories that have the app 
installed, we need them to log into the GitHub app so we can see if they 
can access that installation.

Very helpful for other team members who don't need to install the app 
but need to login to authorize it.