Perform initial sync of subjects using an installation token
Fixes #1125

There was an interesting race condition where the GitHub App was 
installed on a repository and the background job fired before any users 
had authorized the GitHub App oauth.

This meant there wasn't an app_token to download the subjects for 
existing notifications. 

Instead we kick off a sync of all possible subjects for a repository 
using an installation token which doesn't require any users to have 
authorized the GitHub App