Add a prod-ready target to the Dockerfile
This change leverage multi-stage build feature to create build targets
specific to development and production environments. Both targets are
based on a primal target named base. Packages and common gems are
installed in the base target and env-specific gems are installed in
their respective target.

Using the production environment: enables PumaWorkerKiller, pull the
logger level higher (so no more sql queries got dumped) and speed up
the application by enabling some caching.

To minimize the image size, the `build-base` package is now installed
and removed every time `bundle` is run. This effectively reduce the size
of the dev build from 533MB to 364MB (for the dev target). And the
compressed size of the image on Docker Hub goes from 219MB to 159MB.

The impact of both modifications on the build time is negligable:
it gets from 7m57 to 7m47 (on a single build on my computer though),
for the development target. However, the build time for the prod env
is a bit higher: ~9m. That's because the dev target appear first in
the Dockerfile and has to be built first (env if it's not used in
the prod target though).