Document the `GITHUB_APP_JWT` environment variable (#1519)
To properly sync information about the App installation and the resources it has access to, Octobox needs access to a private key for the App. I knew this, since I worked on Apps at GitHub 😂, but this requirement isn't documented for Octobox. I noticed in the code that we have a `GITHUB_APP_JWT` variable (the name seemed a bit off to me, but it's not a big deal), so here's some updated documentation on how to generate the key and use it in Octobox.