Edge case in team review requests
These changes are really tricky, so I've commented a ton in the code
itself.

TLDR, this handles two scenarios:

- Since team_review_request is a fake reason, it'll always get
overwritten by new notification syncs, causing the subject to not be
updated later (there's an intentional delay), and all sorts of issues
therein.
- If someone opens a PR or whatever and requests a team review, _then_
an explicit review from the user, we need to make sure the explicit
review wins.