fix: Add ALLOW_HOSTNAME and ALLOW_ALL_HOSTNAMES environment variables
This is to allow an octobox setup in an reverse proxy setup and to pass
the hostname check as described in https://www.fngtps.com/2019/rails6-blocked-host/

Also, adapted the OpenShift templates to use ALLOW_ALL_HOSTNAMES as
octobox is shielded by the OpenShift router anyway and does not require
this hostname check.

Fixes #2011.