Extract sync instance method
Lots of the stuff going on inside the Subject.sync method is acting on
an instance and not doing things at the class level. This commit extracts
a Subject#sync instance method to make things more object oriented and
to allow us to start extracting the logic out of this method. This
should hopefully make things easier to understand what's going on.